Move the files around in prep for new Strategy.
