+++ examples/legacy-login/package.json
