+++ examples/legacy-login/views/account.ejs
