+++ lib/passport-familysearch/index.js
@@ -1,7 +1,7 @@
 /**
  * Module dependencies.
  */
-var LegacyStrategy = require('./strategy');
+var LegacyStrategy = require('./legacy-strategy');
 
 
 /**
