+++ lib/passport-familysearch/legacy-strategy.js
